Chicken: complete the boundary-layer inflow example
...with some reduction of the flow domain and the addition of
post-processing scripts.

Run time on the GTX 1650 is 48 seconds for 4500 steps.
This gives us a computational cost of
0.194 microseconds per cell per update-stage.
That's 100 times better than the original 2D calculation on a Cray-Y/MP.

We should get another 10 times better on a dxga100 device (according to
the runs that we measured at the GPU hackathon).