eilmer: Added scale_species_after_reconstruction (default=true) to both solvers.
Previously, the transient code would always scale the mass fractions and the
JFNK never would. This change adds an option to control the scaling at run-time
that is needed for high enthalpy multi-species calculations, at the cost of
degraded deep convergence when using the JFNK.

Note that the default=true does change the JFNK's behaviour, which should be
noted when revisiting old simulations.