Lorikeet: fix the second periodic-shear-layer example
... to have its blocks ordered left to right in i-index order.
The order is implicit and needs to be honoured in the way that the
overall grid is stitched together and for the exchange of boundary data.

This test case has now shown that the glitches in vorticity are limited
to the boundaries of the Paraview grid rather than the block boundaries
within the flow solution. Good. Paraview does not know that the flow is
periodic in the x-direction.