eilmer: added input parameter to select which flux calculator is used in the NK solver precondition matrix.
There are some instances, such as when employing the ASF low dissipation flux calculator, when it is recommendable
to use an alternate (typically more dissipative) flux calculator for constructing the precondition matrix.

To choose the precondition matrix flux calculator, set precondition_matrix_flux_calculator (in the steady-state
solver options) to one of the available flux calculator methods. If a flux calculator is not specified for the
precondition matrix, it will default to the main flux calculator.