pitot3: fixed small bug + added ability to not make output files.
Eric Chang at Oxford was having an issue with some of the diaphragm models.
It turns out I didn't check they still work after adding the mole fraction output.
That is fixed now.

I also thought that if Toby is doing some scripting with the code he might appreciate
the ability to have it not make the output files and to remove ALL files created during
the run, so I added the ability to do that too.