kinetics: Fixed bug in initialisation of Arrhenius2
This was a long-standing bug from when I implemented this in 2018 at
Pierre Mariotto's suggestion. Likely, this form of the Arrhenius
expression has never been used by anyone.

In doing this work, I needed to fix a couple of broken tests.
This is why there are changes to files: reaction.d and
reaction_mechanism.d