kinetics: Only use the number density of the collision partner when computing collision frequency in SSH-type relaxation times.
This is contrary to the written descriptions in Rowan's thesis and Thivet's original
paper, but it is what Rowan had in Eilmer3. It also improves agreement
with the Rowan and Thivet's results. This change appears to cause
the separate vibration temperatures to collapse onto the 2T vibration
temperature as the V-V rate is artificially increased (at least for pure
time relaxation problems with no gas-dynamic effects)