nm: Added a transpose method to SMatrix class.
The transpose method creates a new sparse matrix
in the CSR (Compressed Sparse Row) format
(which is the underlying format in SMatrix class).

To do this, we pretend the original matix is
in CSC (Compressed Sparse Column) format and
proceed to convert it to CSR format.

This method will require as much memory as the
original matrix when constructing and returning
the transpose to the caller.