PyGasState class keeps its data directly in Python attributes.
This class is for use in Luke Bartholomew's flow simulation program
that makes frequent use of the gas state's attributes from Python.
The intention is for a PyGasState object to be a drop-in replacement
for a corresponding GasState object.

We have tested only the Brayton cycle example (included in this commit).