lib/gdtk: PyGasState molef setter now only changes Python level massf
We decided (Luke B. and Rowan G.) that the molef setter should
only change the Python-level massf list in a PyGasState object.
We did this for consistency with setting massf at the Python level,
that is, it only has effect at the Python level.
We then rely on a call to any of the update methods to get the
underlying D-GasState mass fractions set.

This is in contrast to the earlier code: setting molef *would* change
the D-level massf but have no effect at the Python level.

The core idea remains that properties at a Python level are not
guaranteed up-to-date until an update function is called.