lib/gdtk: PyGasState now has massf setter and copy_mass_fractions_into_dgs no longer requires type checking.
PyGasState now enables massf to be set using dict or list, then stores
the given data as a list at the PyGasState level. This is for consistency,
where previously, setting massf through a dict would store the massf
property as a dict, whilst GasState converts to a list and stores as a list.

By removing the data type inconsistency from the setter method, this
removed the need to check the data type of the stored massf property
when passing massf to the underlying dlang gas state object when a
state update call is made.