gas: Have negative mass fractions of electrons not contribute anything to the thermodynamics and transport properties in the three temperature model.
I've observed the steady state solver giving negative mass fractions of electrons on the way to steady state (but not necessarily at steady state), which caused stalled convergence. This is due to the electron temperature being poorly defined for small electron mass fractions. Negative mass fractions makes this worse. This commit is a hacky fix for this; it would be nicer to have some other fix.