lmr: Enable setting of reference residuals at step 0
Kyle and Nick had told me it would be useful to allow computing
and setting residuals BEFORE the first step. I didn't like it because
it messed up my code with a special case for getting started. However,
it becomes very important for using the auto-CFL in certain cases
because it gives a repeatable start regardles of initial CFL step.

I have update the current batch of examples to use the reference
residuals set from step 0, and this is the default selection now.