lmr: Add some consistency checks for N-K config.
Specifically, check that:
  1. phase change steps has at least enough entries to
     be consistent with the request for number of phases.
  2. the interpolation orders in any phase do NOT exceed
     the globally-requested interpolation order

Numerically, it makes no sense to go higher order in a phase
than the global request.
Implementation-wise, we size memory based on globally-requested
interpolation order so we can't step to a higher order during a
phase without stopping to do memory allocations.
We won't do this because, as mentioned above, the use case doesn't
make sense from a numerical/algorithm perspective.