lmr: use of local timestep is now controlled at phase level for N-K solver.
The use case for this, as suggested by Kyle, is that sometimes it is
useful to have a globally restrictive timestep in the early phase of
solving some complex flow arrangements.
Later, one can switch to local timestep selection to accelerate
convergence.
This can now be handled in the phase selections. Eg.

NewtonKrylovPhase:new{
   use_local_timestep = false -- sets global dt in this phase
}
NewtonKrylovPhase:new{
   use_local_timestep = true -- now switch to local dt in this phase
}