pitot3: added some things to help Liam do some HYPULSE calculations.
I added a h2/o2/he gas model for Liam to mess around with at Purdue
for HYPULSE calcs.

This required some new species in the species molecular weight list,
and it failed at room temperature so I had to add some code to make that work,
similar to what we do to get CO2 working.