lmr: Change lmr-verify so that a LOGFILE is written for each grid-level...
...and allow "maximum-steps" as a valid reason for stopping. So now,
we consider achieving a relative residual target or stepping
through to maximum steps as successful runs.

The reason for this change is that we might want to set a number for
maximum steps that experience shows gives us machine precision and
run for this number. This should be considered a successful run.