Get Troy's cylinder example running again.
The previous commit got the example up to date with the Lua modules for
the preparation stage.  This commit gets the main simulation running
to completion by changing to a seemingly more robust thermo interpolator
and by tolerating a few bad cells along the way.  Although it runs,
it does not run without complaint by the core solver.  It's a tough
exercise, with a rather harsh expansion over the sharp edge of the
cylinder.