eilmer: shuffled order of entries in conserved quantities vector for Newton-Krlyov solver.
The Newton-Krylov solver had previously used the [1] entry in the U and dUdt vectors as a
place to store perturbed/provisional flow states. This was sufficient for the steady-state
mode since the [0] entry was the only real flow state that needed to be saved. The time-accurate
mode however requires several real flow states to be saved to achieve high-order accuracy.
This revision updates the Newton-Krlyov solver to use the last entry in the U and dUdt vectors
to store perturbed/provisional flow states. This allows several real flow states to be stored
sequentially at the start of the U and dUdt vectors.