eilmer: Disabled the invalidException FloatingPointException flag.
Annoyingly, this exception triggers whenever the code tries to write a NaN to
the terminal, because LDC's write function has to actually test for them using
a comparison operator, which is considered an invalid operation by the hardware.

Given the frequency with which both users and developers print out gas states
and other information for debugging, I believe having an exception thrown
whenever a NaN is present will be a problem.