examples: added an example case for the steady fluid transient solid conjugate heat transfer solver.
The test case is a hollow cylinder subjected to hypersonic flow for 5 seconds.
We will add this test case to the automated list once the CI bot is updated to compile the CHT solver.

This commit also moves the original transient case for this example into a more appropriate location.