lmr: Added gridpro import for prep-grid stage.
A bit of tweaking in this commit for the gridpro import.
In Eilmer4, the gridpro import was focused on getting
to prepped Blocks stage.

In Eilmer5, the focus is different. We try to do all import
of Gridpro grids and their connectivity and BCs at the
stage of grid prep. This commit reflects that change to workflow.