solid: refactor of the residual evaluation routines to operate much as we would an unstructured-grid block.
The catalyst for this refactor was to make the code more readily linearised using our
complex-step differentation approach.
A secondary benefit is that it makes the solid solver code more consistent with how the
flow solver code is structured.
Note that we do take a small performance hit due to the restructuring of the routines.