lmr: Add doNotPerturb status to FVCell and use this in Jacobian formation.
When forming a distributed Jacobian, there are cells in the block which
we'd like to put Jacobian entries in for BUT they themselves we do not
want to perturb. This flag allows us to by-pass doing a perturbation.

There are two reasons we might not want to perturb a cell, one is
algorithmic and the other is implementation. In terms of algorithm.
we may not need the perturbation effect in the Jacobian (since
we'll later transpose it and not use the associated rows).
In terms of implementation, there are cells that don't have full
neighbour information available so doing a perturbation gives
rubbish in the best case, and could lead to trying to access
unavailable data in the worst case.