solid: replaced complex!double type data structures with double type data structures in the Newton-Krylov solver.
This commit removes the (redundant) use of complex numbers from the time-integration data structures since
this data only needs real-valued numbers. The affected data structures include those in the GMRES algorithm.

This brings the solid domain JFNK solver in line with the changes made to the fluid domain variant made in
commit 7832ec564.