geom: Add flexibility to ControlPointPatch -- remove boundary conformity constraint.
In Eiseman's original formulation, he describes how the patch can be
made to conform to boundary shapes, or how it can be left to
follow the control points (near the boundaries). It seemed
natural in our grid generation for Eilmer input scripts that we would always
want to conform to user-supplied boundaries: N, E, S, W. So we had only
coded that option.

However, for joining control point patches together, it might be
convenient to ditch a user-supplied boundary at the joined edges.

In the new implementation, if a boundary is supplied as null
(or nil in Lua), then that boundary has nothing to conform to.
We'll use Eiseman's hints on dropping that term from the tensor
product.

See p. 1172 in:
Eiseman (1988)
A Control Point Form of Algebraic Grid Generation
IJNMF, 8(10):pp. 1165--1181