ntypes: evaluate all integer powers using the integer type routine even if the power is a double type.
Robb Watt has found that the routines for calculating a complex number raised to a double type power,
i.e. pow(complex, double) or complex^^double can be sensitive when the imaginary component is some
small non-zero number, e.g. 10.0+1.0e-70i. This sensitivity is observed even when the double type power
is representable as an integer, in Robb's case it was a power of 6.0. The routines for integer type powers
have been observed to be more robust in these situations. See commit d084311 for more details.

To improve robustness, we now check if a double type power is an integer, if it is, we cast it as an
integer and evaluate the power using the integer type routines. Note that if the power is an integer
but is not representable as an integer type (i.e. the power > int.max) then we just call the double
type routines.