lmr: Added transient mode
This is a large commit. There are a number of new lines of code
in lmr5. There is also a number of files transferred from eilmer4
and reworked. Principally, the reworking involved removing
"base_file_name" from code and replacing how it works.

[TODO] makefile builds real-valued runsim executables.
I'll need to think on how to handle the complex-valued
build for steady mode and the real-valued build for
transient mode separately.