lmr/eilmer: User defined interfaces now ignore mass fractions if no table is provided by the user's lua code.
This bug was found by user gianmarioperrucci on github. Previously the code
would set all mass fractions to zero if the user didn't provide them,
deliberately causing a crash. This interferred with the
WallBC_NoSlip_UserDefinedT boundary condition, which uses the user defined
machinery to set a temperature only, letting other boundary effects deal with
the species values. The old behaviour was also inconsistent with the other
variables in the UDF, which are ignored if not present.