Puffin: fill in code to write a slice of data at given x-location.
Example of use:

$ puffin-post --job=source_test --output=cross --x-location=1.05
$ gnuplot
gnuplot> plot "source_test/source_test-cross-1.txt" using 2:5

should result in the Mach number being plotted against y-location
of the cells in streamtube 1 for Ingo's source_test exercise.
The cross-stream data files are in a text format that GNU Plot can read,
with the first line being a comment giving the column names.
If the x-location is not found, no data values will be written.