pitot3: fixed tiny bug in species MW import
I moved the species MW dict (which we use to calculate mole fractions)
to a function and now it seems to remember something about how the
YAML load is changed, so PITOT3 fails on a second scripted run.

Weird issue, but fixed now by just not doing stuff if it sees
that they don't need to be done.