lmr: specialise the checks for the unstructured sharp cone based on OS
Linux and MacOS have a slightly different journey to convergence on the
sharp cone case in unstructured grid mode. The reason is likely down
at the level of floating point precision. However, I think there might
be some changes we can make with algorithm use that reduces the sensitivity.
I'll talk to Kyle about that.

In the meantime, the test script acknowledges this difference in convergence
journey and tests based on OS. I check the flow field output for both linux
and mac; they give a converged cone20 flow field that looks perfectly fine to me.