lmr: Re-connect the reading of grids when grid_motion != none in FlowSolution
I had a placeholder for a second constructor in FlowSolution
that would be specialise for transient simulations.
It turns out we don't need that, so I've removed it.
The only special action is to test on grid_motion.

(and maybe one day the steady-state solver will do shock-fitting,
so grid motion will be fair game.)