lmr: Rename FVCell --> FluidFVCell
This commit touches many files.

This rename is to allow abstraction (*) at the block I/O level.
The block I/O arrangement doesn't really care about the type
of a "Cell". It just wants to access data (as provisioned by the cell)
and write to disk.

(And removed gas_solid_interface file. The entire contents were
commented out. It was doing nothing for us.)

(*) To appear in a later commit.