lmr: disable floating point exceptions for complex number version compiled with optimizations...
When compiling in fast mode (e.g. with -O1, -O2, etc.) the LDC2 compiler
optimizes the code to pre-calculate both branches of conditional statements.
This can cause problems when floating point exceptions are enabled, since
one of the branches, the invalid or 'forbidden' branch, may peform an invalid
floating point operation, such as division by zero. We have observed that
the routine in the complex number module which handles complex number division
exhibits this behaviour. This does not occur for the debug version of the code
since no optimizations are enabled, and thus only the 'safe' branch of the code
is executed. For this reason we only switch on floating point exceptions for
the debug flavour of the complex number version.