lmr & examples/lmr: introduction of FVCell interface and name changes.
This is a large commit and achieves several things because they were
so intertwined. The goal is to prepare the code for CHT simulations
and the associated I/O for solidblocks.

An FVCell interface was introduced as the abstract base class
for both FluidFVCell and SolidFVCell.

Since we started to add the qualifier Fluid to the cells this
propagated into the I/O.

Then I got to the point that "flow_format" didn't really make
sense for the solid domain and I didn't see a good case for
having separate formats for the fluid and solid files on disk
in a single simulation (but I could think of many reasons
of why it was a bad idea). I unified this concept into "field_format".