lmr: light refactoring of linear system scaling for Newton-Krylov solver.
We now store the row and column scaling factors separately.

This commit should have no functional change to field solutions obtained
using the Newton-Krylov solver, however, there might be small changes to the
convergence history due to changes in the order of floating point operations.

Modified the expected number of steps and CFL for the steady-state version
of the 2D/sharp-cone-20-degrees/usg-minimal pytest.