examples/lmr: Similar adjustments in cone/sg-minimal on CFL growth as per previous commit beafee0e
Drop the rate of CFL growth seems to make this case less sensitive.

I increased the number of first-phase steps also.

We actually get to convergence in fewer steps (now 60). So this
seems to be a case of taking a slower start and more first-order
steps accelerates the convergence when we approach the end game.

Tests adjusted and flow field output was checked.