lmr: Disable shared-memory parallelsim in Newton-Krylov solver.
The new scaling arrangement was not multi-thread compatible and
adding the 'shared' qualifier on a ConservedQuantities array caused
a mess that could not easily be unravelled.

The solution for now is to disable the shared-memory parallelism
in the Newton-Krylov solver. The MPI parallelism is working
nicely and scales well on both multi-core and multi-node machines.