lmr: added residuals to optional output data.
The Newton-Krylov solver can now write out the cell residuals to disk by setting
write_residual_values to true in the NewtonKrylovGlobalConfig table.

This commit also adds a new command as well, residual2vtk, which converts the residual
values stored in native format to VTK files which can be visualized in Paraview.