lmr: grand unification of the loads output in eilmer5.
This unifies the loads output for the steady and transient solver modes.

It also unifies the loads output with the snapshots output, i.e. the loads
directory now follows a similar structure as the snapshots directory.

The subdirectories are named by the loads index (we have dropped the "t" prefix)
and the information regarding what the sim_time and step are for the corresponding
loads subdirectory is stored in the loads-metadata file (renamed from loads.times).