lmr-examples: modified the Hakkinen steady-state example to improve performance.
The structured-grid limiter appears to be noisy for this simulation, this is
causing the Newton-Krylov solver to require an inordinate number of inner
iterations to converge each linear system.

To alleviate this issue, I have switched off the base limiter and switched
on the heuristic pressure based limiter, which appears to be less problematic
for this flow field.