lmr: transient code now writes out a notification when loads are being written to disk.
I took the opportunity to also package up the sequence of routine calls
to write the loads to disk for the transient code into a function, similar
to the steady solver code.

This commit also makes some minor cosmetic changes to the loads and
snapshots notifications for both the transient and steady solvers.