pitot3: driver condition and test gas selections now try to help a bit
now if the user tries to specify a driver condition and gets it wrong
the code will print all of the driver conditions for that facility or
the files in the folder if it is a custom driver condition.

It does a similar thing for pre-set and custom gas models too.