lmr: Consolidate how MPI build is handled in makefile
I took Nick's suggestion of consolidating MPI executables regardless
of OpenMPI choice or MPICH.

The choice of MPI implementation when building is an either/or deal:
either OpenMPI (default) or MPICH (WITH_MPICH=1). We have the MPICH
option for building on the Western Australian cray supercomputers.

OPT_OUT_MPI=1 is still available when no MPI is required.