gas: Initialise the number of modes of multi-temperature gas before setting up the reverse energy mode lookup table
Setting up the reverse energy mode lookup table requires the number of energy modes to be set properly.
For the multi-temperature gas, the number of modes was being set after the table was built, so the table wasn't built properly.
Eilmer 4 didn't use the reverse energy mode lookup table, so this didn't really matter.
But Eilmer 5's IO uses the table, so the energy modes weren't being set properly during the preparation stage of a simulation