lmr: added an option to Newton-Krylov solver to enforce the linear solver relative residual tolerance.
Sometimes it is favourable to enforce the solver to only accept nonlinear updates that satisfy the
linear solver relative residual tolerance. Although this might result in a longer wall-clock time
to finish for unproblematic simulations, it can provide some robustness in the early phases of a
problematic simulation.

The default is set to false, to turn on for a particular phase, set the following in the phase settings:

    enforce_linear_solver_tolerance = true

If the linear solver tolerance is not met during a linear system solve, the nonlinear update is discarded
and the nonlinear step is attempted again with a reduced CFL.