lmr: fixed a bug in the code that reads in a previous flow solution stored in binary format.
The functions that generate the flow and grid filenames require that the field_format and
grid_format variables attached to the GlobalConfig object have been correctly set.

However, when prepping a simulation, the GlobalConfig object is not guaranteed to have been filled.

We now set these variables in the GlobalConfig object as soon as we read them in from the
config file in the FlowSolution and SolidSolution constructors.