lmr: fixed a bug in the Zones functionality (e.g. turbulentZones).
The bug fix required adding the Zones table definitions to prepgrids.lua.

Although prep-grids does not need information regarding Zones, it will try operate on
these tables if a Zone is defined in the same script that the grid is defined -
i.e. when the grid and simulation are defined in one monolithic Lua input script.