lmr: Change to solid domain infrastructure for state/properties calculation.
This commit introduces a SolidThermalModel that is attached
to each solid domain block. We store this per block so that
various materials can be present in a simulation. The follow-on
restriction is that blocks must only contain a single physical
material. If there is a material interface (say nickel to steel),
then there will need to be a block boundary.

There is only one SolidThermalModel presently: a constant properties
model. This should replicate the behaviour prior to this commit.

Next step is to add a model where properties of materials
vary linearly with temperature.

This commit obviated the need for a solidprops.d module.
It has been removed.

MAINTENANCE NOTE:
I removed the anisotropic properties for solids in this commit.
I think that code path has been untested for some time,
it's not verified, and it's not part of the tests.