lmr: Bug fix for picking up times in post-processing.
For a single snapshot to process, the time (if SolverMode.transient)
was not being picked up correctly because the code defaulted
to assuming we were working with an initial flow field. Now fixed.

This also means that we can correctly pass through the time to the
compute-norms command for time-dependent reference solutions.