lmr: Solid domain passing first tests for linearly-varying properties.
Some additions in this commit:

1. e-T relationship with Cp as linear function of T tested.
2. thermal conductivity at interfaces is updated just
   after taking an average temperature at interfaces.

TODO: We need to make the e-T relationship handle the cases when
we step off the linear range. Out thinking is that we will
assume that thermal properties are constant off the ends
of the linear range.
(If a user does not like that, they can make their linear range
bigger. Then they can make the decision about properties going
negative if the line doesn't have a nice intercept with zero.)