lmr/solid: removed the energy positivity check for the solid domain solver.
With the addition of temperature dependent thermal properties, the solver may
now have valid negative internal energy, depending on what is chosen as the
reference energy.

This removal is temporary to allow Saxon to perform some bounds checking on
the new implementation while we construct a more appropriate test for
non-physical solid domain states.