lmr: added functionality to skip reading in specific variables from a snapshot file.
We identified an inconsistency in setting the geometric properties during start-up.

During the init stage of a simulation, all geometric properties were being computed from
the grid vertex positions. After this had occured, the flowfield was being initialised by
reading in data from the snapshot files (note that the snapshot files also contained cell
centroid data, e.g. pos.x, pos.y, pos.z).

This meant that if the grid files and the snapshot files were not consistent, then the geometric
cell centroids were not consistent with the other geometric properties (vertex positions, cell
volumes, face areas, etc.).

We can now pass a list of variable names to skip when reading in snapshot data. We use this
functionality to skip reading in the cell position data during start-up. This ensures that
all geometric properties are consistent with the current grid.