lmr: Add a more nuanced implementation for update_thermo_trans_coeffs boundary interface effect.
There was an inconsistency between setting a thermo_interpolator option
and how the update_thermo_trans_coeffs boundary interface effect worked.
The latter *always* did a pT update regardles of what was occuring at
all internal interfaces.

In this commit, I've added a parameter to the update_thermo_trans_coeffs
implementation that takes an option for thermo interpolator. It isn't
really intended for uses to control It is configured in the back end.
The configuration is:

1. Use the globally configured thermo_interpolator (for internal
   interfaces) at internal exchange BCs.

2. Use the "pT" interpolator at wall-type BCs to preserve current
   behaviuor near walls and because setting temperature at a wall
   is a fairly common no-slip BC.

The "pT" option mimics old behaviour because that was the only
thermo update being called.