lmr: Added a solid thermal model that works from tabulated properties.
This compiles, not tested. That's a job for Saxon James.

The tabulated properties model works on doing a linear variation
of properties BETWEEN table entries. The table is specified as a plain
text file with comma separated values as:

T,     k,      Cp,     e
<T0>,  <k0>,   <Cp0>,  <e0>
<T1>,  <k1>,   <Cp1>,  <e1>
etc....

where actual (float) values replace symbols in < >.

The user also specifies density for the solid. This is assumed constant
at the specified value.