pitot3: updated some RST things
It turns out that overall my RST implementation into PITOT3 was pretty bad.
It didn't output the reflected shock speed in the output for example,
that is now fixed, and a lot of detail about that calculation
was not printed during the run, which is now fixed too.

I think there was also an issue with the reflected shock through the driver
gas where it was being stagnated too instead of just shocked by the
same shock speed reflected shock as the test gas. I have changed this so it is
shocked by the same shock as the reflected test gas. I did a lot of reading,
and I think this is right, but I will check with Richard and update if needed.