lib/gdtk: Reverse the sample points in the Busemann contour
In the integration process, the Busemann contour is built
upstream in what we typically consider a negative x direction.

For using the contour, it's more useful to have points ordered
in the streamwise direction from upstream towards downstream.
In this commit, I have reversed the points when writing to files
and building splines.