examples/lmr: Add the Busemann diffuser example to Eilmer.
This example mimics the same configuration as the diffuser
example in Puffin. I have put that note in the README.

The simulation makes use of manual grid sequencing. The user
is expected to run the coarse-grid before proceeding to
run the fine grid.

A first-pass version of the README is added, but needs some
fleshing out.