lmr: fixed a bug in the Lua code for generating the nk-config file.
Replaced x = x or v Lua idiom with explicit check on whether x is nil.
The problem with using x = x or v is that x will be set to v if x is nil OR false (see https://www.lua.org/pil/3.3.html).
In this use case, x, which is a boolean user input parameter, may very well be defined as false.
In this instance, x will not be set correctly in the nk-config file if v (the default value) is defined as true.