lmr: fixed a bug in the physicality check.
After determining the minimum omega across all blocks based on
the allowable mass change, we want to then transmit this global
minimum back to the blocks before checking if the primitive variables
are valid (i.e. produce a realizable thermodynamic state). This could
reduce the number of costly decode_conserved() routine calls in the
latter stage.