lmr: added gradients (both convective and viscous) to optional output data.
The Newton-Krylov solver can now write out the cell gradients (if they are available) to disk by
setting write_gradient_values to true in the NewtonKrylovGlobalConfig table.

This commit also adds a new command as well, gradient2vtk, which converts the gradient
values stored in native format to VTK files which can be visualized in Paraview.

Note that this data isn't really necessary for production simulations, and is intended to be
used only by developers.