examples/lmr: Add the diffuser-busemann to automated tests.
Note this is a two-for-one deal. The test exercises the fine-grid
case for the simulation. The fine-grid simulation happens to be
warm-started from the coarse-grid sim. So that is also tested
as a by-product of running this new test.

I have orchestrated the testing from the fine-grid directory
because I want to test both the result of the fine-grid simulation
and the workflow sequence to get there.

Additionally, I made the makefile a little smarter about deciding
between command names 'pytest' and 'pytest-3'.