lmr: Add proper handling of duplicate history cell requests.
Sometimes a user might request duplicate history cell locations.
When this occurs, we proceed with the principle of "you get what you ask for".
The code will produce two separate history files, even if they contain the
same information.

The user is given a warning about this so they are not surprised. An example
is here:

   W: Duplicate history cell requested.
   W: Earlier cell is: 01
   W: Current request is: 03
   W: Cell is [blk:idx]= [1:9] located at: Vector3(0.451313, 0.101727, 0)
   W: At run-time, the code will produce one file per history cell request.
   W: For this case, you will receive two files with identical information.