lmr: implemented more appropriate handling of unstructured inviscid gradients in ghost cells.
Originally we had been copying the internal cell's least-squares gradient into the ghost cell for the unstructured grid solver for all boundary domain ghost cells (e.g. non-mapped cell boundary conditions). This is appropriate for several of our boundary conditions, however, for "wall" boundary conditions it is more appropriate to reflect the wall-normal component of the gradient (similar to how we reflect the velocity along such boundary conditions).

This commit now performs this gradient reflection along "wall" boundary conditions. To reuse some of the Vector3 transformation routines, we have converted all of the least-squares gradient storage types in lsqinterp.d to Vector3.

This commit slightly refactors the convective flux code path for the unstructured solver. The new logic is as follows:

convective_flux_phase0 - this routine computes the least-squares gradients in a block, it then finds any non-mapped cell boundary conditions and copies the internal cell gradient into the ghost cell, finally it then finds any "wall" boundary conditions and reflects the wall-normal component of the gradients in the ghost cells.

convective_flux_phase1 - this routine computes the gradient limiter values in a block, it then finds any non-mapped cell boundary conditions and copies the internal cell limiter values into the ghost cells.

convective_flux_phase2 - this routine reconstructs the left and right flow states and performs the flux evaluation.

Note that mapped cell communication is performed between phase0 and phase1 to transfer up-to-date gradients, and between phase1 and phase2 to transfer up-to-date limiter values.