lmr: allow user to set the limit on cfl increase/decrease ratio for the Newton-Krylov solver.
We currently have these set to default values of 2.0 and 0.1, respectively.

These defaults have been sufficient for a broad range of simulations.

However, I have observed that for a couple of simulations, where the global residual doesn't drop monotonically,
the CFL can get stuck in some small value (e.g. < 1.0). For these cases, it is helpful to select more
appropriate values for the limit on cfl increase/decrease ratio to improve convergence.