lmr: refactored the Park limiter to follow the original algorithm more closely.
The implementation of the Park limiter was initially a modification on the algorithm
presented in Park's thesis. This was because at the place of the limiter evaluation in
the solver, the gradients in domain boundary ghost cells were not up to date.

Now that the ghost cell gradients are filled at the time of the interior cells (as of
commit a9c270f), we can implement the original algorithm as defined in Park's thesis.

The effect of the change is that the limiter is stronger in regions with large flow field
gradients and is weaker in regions where the flow field variables are smoothly varying.

Note that we still apply the harmonic modification implemented by NNG.