lmr: Makefile now always builds an 'lmr-debug' executable.
An lmr-debug executable is always built.

For the other executables (lmr, lmr-run, etc.), the behvaiour is unchanged.
If you request FLAVOUR=fast, those variants are built.
If you request FLAVOUR=debug, those variants are built.

The reason for this change will become apparent in the next commit.
We wanted to be able to call the debug executable when doing preparation
involving the CEAGas model.