lmr: implemented a least-squares gradient evaluation variant based on QR factorization.
The literature suggests that solving the normal equations via a QR factorization is more stable
than inverting the normal equations directly, as we currently do. In particular, QR factorization
is said to be superior for highly stretched grids with curvature.

I have implemented the QR factorization method to explore this assertion.

NOTE: the default behaviour is still to use unweighted least-squares and weighted least-squares
      solved via inverting the normal equations directly. QR factorization is not intended
      for wider use at this stage.