examples/lmr: added steady-state solver variants of the sphere-heat-transfer example from examples/eilmer.
There are two cases in this example: one which uses a natively generated quad-based unstructured grid; and
one which provides an example of how to generate an unstructured triangluar mesh with a quad-based prism layer
in GMSH for use with Eilmer.

The two cases have also been added to the suite of automated tests.