example/lmr: added a pre-generated su2 grid for the sphere-heat-transfer GMSH example.
The autobot operates on Rangpur which has GMSH 4.11 installed. This appears to generate a grid
that is substantially different to GMSH 4.13, which is installed on my local machine. This difference
leads to a test failure for the sphere-heat-transfer case.

We have decided to commit the generated su2 grid for the automated test for this reason.

The GMSH generation script is still remaining as an example for users.